+++ src/pages/result/components/sent-view.tsx
